+++ auth/presentation/src/main/kotlin/me/proton/core/auth/presentation/ui/Utils.kt
@@ -21,13 +21,16 @@ package me.proton.core.auth.presentation.ui
 import androidx.activity.ComponentActivity
 import androidx.fragment.app.FragmentManager
 import me.proton.core.auth.presentation.R
+import me.proton.core.auth.presentation.ui.signup.SignupFinishedFragment
 import me.proton.core.auth.presentation.ui.signup.CreatingUserFragment
+import me.proton.core.domain.entity.Product
 import me.proton.core.presentation.ui.alert.ProtonCancellableAlertDialog
 import me.proton.core.presentation.utils.inTransaction
 import me.proton.core.presentation.utils.openBrowserLink
 
 private const val TAG_PASSWORD_CHOOSER_DIALOG = "password_chooser_dialog"
 private const val TAG_CREATING_USER = "creating_user_fragment"
+private const val TAG_CONGRATS = "congrats_fragment"
 
 /**
  * Presents to the user a dialog to inform that it should update the password inn order to be able to login.
@@ -77,6 +80,28 @@ fun FragmentManager.showCreatingUser(
 
 fun FragmentManager.removeCreatingUser() {
     findFragmentByTag(TAG_CREATING_USER)?.let {
+        inTransaction {
+            setCustomAnimations(0, 0)
+            remove(it)
+        }
+        popBackStack()
+    }
+}
+
+fun FragmentManager.showCongrats(
+    containerId: Int = android.R.id.content
+) = findFragmentByTag(TAG_CONGRATS) ?: run {
+    val congratsFragment = SignupFinishedFragment()
+    inTransaction {
+        setCustomAnimations(0, 0)
+        add(containerId, congratsFragment, TAG_CONGRATS)
+        addToBackStack(null)
+    }
+    congratsFragment
+}
+
+fun FragmentManager.removeCongrats() {
+    findFragmentByTag(TAG_CONGRATS)?.let {
         inTransaction {
             setCustomAnimations(0, 0)
             remove(it)