+++ coreexample/src/main/kotlin/me/proton/android/core/coreexample/db/AppDatabaseMigrations.kt
@@ -122,4 +122,10 @@ object AppDatabaseMigrations {
             FeatureFlagDatabase.MIGRATION_0.migrate(database)
         }
     }
+
+    val MIGRATION_14_15 = object : Migration(14, 15) {
+        override fun migrate(database: SupportSQLiteDatabase) {
+            OrganizationDatabase.MIGRATION_1.migrate(database)
+        }
+    }
 }