+++ test/android/instrumented/src/main/kotlin/me/proton/core/test/android/plugins/data/Plan.kt
@@ -24,7 +24,7 @@ import me.proton.core.test.android.instrumented.ProtonTest.Companion.getTargetCo
 
 @Serializable
 enum class Plan(var planName: String, var text: String) {
-    Free("free", "Free"),
+    Free("free", "Proton Free"),
     Professional("pro", "ProtonMail Professional"),
     Visionary("visionary", "Visionary"),
     Plus("plus", "ProtonMail Plus"),